requirements.txt
torch>=2.0.0,<2.1.0
transformers>=4.30.0,<4.31.0
numpy>=1.24.0,<1.27.0
scipy>=1.10.0,<1.11.0
scikit-learn>=1.3.0,<1.4.0
matplotlib>=3.7.0,<3.8.0
seaborn>=0.12.0,<0.13.0
tqdm>=4.65.0,<4.66.0
wandb>=0.15.0,<0.16.0
tensorboard>=2.13.0,<2.14.0
einops>=0.6.0,<0.7.0
optuna>=3.2.0,<3.3.0
datasets>=2.12.0,<2.13.0
accelerate>=0.20.0,<0.21.0
deepspeed>=0.9.0,<0.10.0
bitsandbytes>=0.39.0,<0.40.0
pyyaml>=6.0,<7.0
h5py>=3.8.0,<3.9.0
psutil>=5.9.0,<6.0.0
flask>=2.3.0,<3.0.0
flask-cors>=4.0.0,<5.0.0
flask-socketio>=5.3.0,<6.0.0
matplotlib>=3.7.0,<3.8.0
pandas>=2.0.0,<3.0.0
pynvml>=11.4.0,<12.0.0
scipy>=1.10.0,<1.11.0
statsmodels>=0.14.0,<0.15.0
bcrypt>=4.0.0,<5.0.0
PyJWT>=2.8.0,<3.0.0
optuna>=3.2.0,<3.3.0
